requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.2
django-admin-interface==0.16.1
django-colorfield==0.4.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-otp==1.0.3
idna==2.10
pyTelegramBotAPI==3.7.7
pytz==2021.1
qrcode==6.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
